Move and rename files associated with importing/exporting/testing stockflow_ascet
