+++ tests/test_stockflow_ascet_source.py
@@ -1,6 +1,7 @@
 from copy import deepcopy as _d
-from mira.sources.stock_and_flow import *
+from mira.sources.askenet.stockflow_ascet import *
 import requests
+
 
 def set_up_file():
     return requests.get(
